notifications

+++ src/app/dashboard/schedules/info-schedule/info-schedule.component.ts
@@ -1,8 +1,10 @@
 import { Component, Inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatDialogRef, MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
 import { HttpClient } from '@angular/common/http';
 import { CreateScheduleComponent } from '../create-schedule/create-schedule.component';
 import { TimeSchedule } from 'src/app/models/TimeSchedule';
+import { SchedulesService } from 'src/app/services/schedules.service';
+import { ConfirmationDialogComponent } from 'src/app/utilities/confirmation-dialog/Confirmation-Dialog.component';
 
 @Component({
   selector: 'info-schedule',
@@ -23,21 +25,25 @@ export class InfoScheduleComponent{
   days: string;
   schoolClassYear: string;
   tipoAula: string;
+  functionType: string
 
 
   constructor(
     public dialogRef: MatDialogRef<CreateScheduleComponent>,
+    public dialog: MatDialog,
     private http: HttpClient,
+    private schedulesService: SchedulesService,
     @Inject(MAT_DIALOG_DATA) public data: any
   ) {
 
 
 
-    ({id: this.id} = data);
+    ({id: this.id, functionType: this.functionType} = data);
 
 
-    this.http.get<TimeSchedule>('http://localhost:3000/time-schedule/' + this.id)
+    this.schedulesService.getTimeSchedule(this.id)
     .subscribe((timeSchedule: TimeSchedule) => {
+      console.log(timeSchedule)
       this.title = timeSchedule.curricularUnit.name;
 
       this.classroom = timeSchedule.classroom.block + "." + timeSchedule.classroom.floor + "." + timeSchedule.classroom.classroomNumber
@@ -56,17 +62,29 @@ export class InfoScheduleComponent{
 
   }
 
+  isSecretariado(): boolean{
+    return this.functionType == "secretariado"
+  }
+
   onNoClick(){
     this.dialogRef.close("cancelou");
   }
-
   onDelete(){
-    this.http.delete(`http://localhost:3000/time-schedule/${this.id}`)
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem a certeza que deseja excluir este horário?',
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.http.delete(`http://localhost:3000/time-schedule/${this.id}`)
   .subscribe(
     () => console.log('Registro removido com sucesso'),
     error => console.log('Ocorreu um erro ao remover o registro:', error)
   );
     this.dialogRef.close("eliminado");
+      }
+    });
+
   }
 
   onEdit() {
